src/requirements.txt
cachetools==3.1.1
certifi==2019.9.11
chardet==3.0.4
defusedxml==0.6.0
Django==2.2.24
django-allauth==0.40.0
django-filter==2.2.0
django-rest-auth==0.9.5
django-storages==1.7.2
djangorestframework==3.11.2
enum34==1.1.6
google-api-core==1.14.3
google-auth==1.30
google-cloud-core==1.0.3
google-cloud-storage==1.23.0
google-resumable-media==0.5.0
googleapis-common-protos==1.6.0
idna==2.8
Markdown==3.1.1
oauthlib==3.1.0
protobuf==3.10.0
psycopg2==2.8.4
pyasn1==0.4.7
pyasn1-modules==0.2.7
python3-openid==3.1.0
pytz==2019.3
requests==2.26.0
requests-oauthlib==1.2.0
rsa==4.7
six==1.12.0
sqlparse==0.3.0
urllib3==1.26.7